Add method to load ROM from byte array

+++ DotNetNdsToolkit/NdsRom.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Collections.Concurrent;
+using SkyEditor.Core.TestComponents;
 
 namespace DotNetNdsToolkit
 {
@@ -659,6 +660,68 @@ namespace DotNetNdsToolkit
         /// </summary>
         public string DataPath { get; set; }
 
+        /// <summary>
+        /// Reads
+        /// </summary>
+        private async Task LoadRomHeader()
+        {          
+            Header = new NdsHeader();
+            Header.CreateFile(await ReadAsync(0, 512));
+
+            var loadingTasks = new List<Task>();
+
+            // Load Arm9 Overlays
+            var arm9overlayTask = Task.Run(async () => Arm9OverlayTable = await ParseArm9OverlayTable());
+
+            if (IsThreadSafe)
+            {
+                loadingTasks.Add(arm9overlayTask);
+            }
+            else
+            {
+                await arm9overlayTask;
+            }
+
+            // Load Arm7 Overlays
+            var arm7overlayTask = Task.Run(async () => Arm7OverlayTable = await ParseArm7OverlayTable());
+
+            if (IsThreadSafe)
+            {
+                loadingTasks.Add(arm7overlayTask);
+            }
+            else
+            {
+                await arm7overlayTask;
+            }
+
+            // Load FAT
+            var fatTask = Task.Run(async () => FAT = await ParseFAT());
+
+            if (IsThreadSafe)
+            {
+                loadingTasks.Add(fatTask);
+            }
+            else
+            {
+                await fatTask;
+            }
+
+            // Load FNT
+            var fntTask = Task.Run(async () => FNT = await ParseFNT());
+
+            if (IsThreadSafe)
+            {
+                loadingTasks.Add(fntTask);
+            }
+            else
+            {
+                await fntTask;
+            }
+
+            // Wait for all loading
+            await Task.WhenAll(loadingTasks);
+        }
+
         public override async Task OpenFile(string filename, IIOProvider provider)
         {
             if (provider.FileExists(filename))
@@ -673,63 +736,7 @@ namespace DotNetNdsToolkit
 
                 VirtualPath = provider.GetTempDirectory();
 
-
-                // Load data
-                Header = new NdsHeader();
-                Header.CreateFile(await ReadAsync(0, 512));
-
-                var loadingTasks = new List<Task>();
-
-                // Load Arm9 Overlays
-                var arm9overlayTask = Task.Run(async () => Arm9OverlayTable = await ParseArm9OverlayTable());
-
-                if (IsThreadSafe)
-                {
-                    loadingTasks.Add(arm9overlayTask);
-                }
-                else
-                {
-                    await arm9overlayTask;
-                }
-
-                // Load Arm7 Overlays
-                var arm7overlayTask = Task.Run(async () => Arm7OverlayTable = await ParseArm7OverlayTable());
-
-                if (IsThreadSafe)
-                {
-                    loadingTasks.Add(arm7overlayTask);
-                }
-                else
-                {
-                    await arm7overlayTask;
-                }
-
-                // Load FAT
-                var fatTask = Task.Run(async () => FAT = await ParseFAT());
-
-                if (IsThreadSafe)
-                {
-                    loadingTasks.Add(fatTask);
-                }
-                else
-                {
-                    await fatTask;
-                }
-
-                // Load FNT
-                var fntTask = Task.Run(async () => FNT = await ParseFNT());
-
-                if (IsThreadSafe)
-                {
-                    loadingTasks.Add(fntTask);
-                }
-                else
-                {
-                    await fntTask;
-                }
-
-                // Wait for all loading
-                await Task.WhenAll(loadingTasks);
+                await LoadRomHeader();
             }
             else if (provider.DirectoryExists(filename))
             {
@@ -742,6 +749,13 @@ namespace DotNetNdsToolkit
             {
                 throw new FileNotFoundException("Could not find file or directory at the given path", filename);
             }
+        }
+
+        public async Task OpenFileInMemory(byte[] rawData)
+        {
+            base.CreateFile(rawData);
+            this.CurrentIOProvider = new MemoryIOProvider();
+            await LoadRomHeader();
         }
 
         public async Task<bool> IsOfType(GenericFile file)
@@ -1645,7 +1659,8 @@ namespace DotNetNdsToolkit
 
         bool IIOProvider.FileExists(string filename)
         {
-            return CurrentIOProvider.FileExists(GetVirtualPath(filename)) || GetFATEntry(filename, false).HasValue;
+            return (CurrentIOProvider != null && CurrentIOProvider.FileExists(GetVirtualPath(filename)))
+                || GetFATEntry(filename, false).HasValue;
         }
 
         private bool DirectoryExists(string[] parts)
@@ -1687,7 +1702,11 @@ namespace DotNetNdsToolkit
 
         bool IIOProvider.DirectoryExists(string path)
         {
-            return !BlacklistedPaths.Contains(FixPath(path)) && (CurrentIOProvider.DirectoryExists(GetVirtualPath(path)) || DirectoryExists(GetPathParts(path)));
+            return !BlacklistedPaths.Contains(FixPath(path))
+                    &&
+                    ((CurrentIOProvider != null && CurrentIOProvider.DirectoryExists(GetVirtualPath(path)))
+                        || DirectoryExists(GetPathParts(path))
+                    );
         }
 
         void IIOProvider.CreateDirectory(string path)
@@ -1700,7 +1719,7 @@ namespace DotNetNdsToolkit
 
             if (!(this as IIOProvider).DirectoryExists(fixedPath))
             {
-                CurrentIOProvider.CreateDirectory(GetVirtualPath(fixedPath));
+                CurrentIOProvider?.CreateDirectory(GetVirtualPath(fixedPath));
             }
         }
 
@@ -1761,7 +1780,7 @@ namespace DotNetNdsToolkit
 
                     // Apply shadowed files
                     var virtualPath = GetVirtualPath(parts[0].ToLower());
-                    if (CurrentIOProvider.DirectoryExists(virtualPath))
+                    if (CurrentIOProvider != null && CurrentIOProvider.DirectoryExists(virtualPath))
                     {
                         foreach (var item in CurrentIOProvider.GetFiles(virtualPath, "overlay_*.bin", true))
                         {
@@ -1792,7 +1811,7 @@ namespace DotNetNdsToolkit
 
                     // Apply shadowed files
                     var virtualPath7 = GetVirtualPath(parts[0].ToLower());
-                    if (CurrentIOProvider.DirectoryExists(virtualPath7))
+                    if (CurrentIOProvider != null && CurrentIOProvider.DirectoryExists(virtualPath7))
                     {
                         foreach (var item in CurrentIOProvider.GetFiles(virtualPath7, "overlay_*.bin", true))
                         {
@@ -1836,7 +1855,7 @@ namespace DotNetNdsToolkit
 
                         // Apply shadowed files
                         var virtualPathData = GetVirtualPath(path);
-                        if (CurrentIOProvider.DirectoryExists(virtualPathData))
+                        if (CurrentIOProvider != null && CurrentIOProvider.DirectoryExists(virtualPathData))
                         {
                             foreach (var item in CurrentIOProvider.GetFiles(virtualPathData, searchPattern, topDirectoryOnly))
                             {
@@ -1889,7 +1908,7 @@ namespace DotNetNdsToolkit
 
                         // Apply shadowed files
                         var virtualPathData = GetVirtualPath(path);
-                        if (CurrentIOProvider.DirectoryExists(virtualPathData))
+                        if (CurrentIOProvider != null && CurrentIOProvider.DirectoryExists(virtualPathData))
                         {
                             foreach (var item in CurrentIOProvider.GetDirectories(virtualPathData, topDirectoryOnly))
                             {
@@ -1927,7 +1946,7 @@ namespace DotNetNdsToolkit
             else
             {
                 var virtualPath = GetVirtualPath(fixedPath);
-                if (CurrentIOProvider.FileExists(virtualPath))
+                if (CurrentIOProvider != null && CurrentIOProvider.FileExists(virtualPath))
                 {
                     return CurrentIOProvider.ReadAllBytes(virtualPath);
                 }
@@ -1952,7 +1971,7 @@ namespace DotNetNdsToolkit
                 BlacklistedPaths.Remove(fixedPath);
             }
 
-            CurrentIOProvider.WriteAllBytes(GetVirtualPath(filename), data);
+            CurrentIOProvider?.WriteAllBytes(GetVirtualPath(filename), data);
         }
 
         void IIOProvider.WriteAllText(string filename, string data)
@@ -1974,7 +1993,7 @@ namespace DotNetNdsToolkit
             }
 
             var virtualPath = GetVirtualPath(filename);
-            if (CurrentIOProvider.FileExists(virtualPath))
+            if (CurrentIOProvider != null && CurrentIOProvider.FileExists(virtualPath))
             {
                 CurrentIOProvider.DeleteFile(virtualPath);
             }
@@ -1989,7 +2008,7 @@ namespace DotNetNdsToolkit
             }
 
             var virtualPath = GetVirtualPath(path);
-            if (CurrentIOProvider.FileExists(virtualPath))
+            if (CurrentIOProvider != null && CurrentIOProvider.FileExists(virtualPath))
             {
                 CurrentIOProvider.DeleteFile(virtualPath);
             }
@@ -2009,8 +2028,13 @@ namespace DotNetNdsToolkit
             return path;
         }
 
-        private Stream OpenFile(string filename, FileAccess access)
+        Stream IIOProvider.OpenFile(string filename)
         {
+            if (CurrentIOProvider != null)
+            {
+                throw new NotSupportedException("Cannot open a file as a stream without an IO provider.");
+            }
+
             var virtualPath = GetVirtualPath(filename);
             if (!CurrentIOProvider.DirectoryExists(virtualPath))
             {
@@ -2018,22 +2042,41 @@ namespace DotNetNdsToolkit
             }
             CurrentIOProvider.WriteAllBytes(virtualPath, (this as IIOProvider).ReadAllBytes(filename));
 
-            return File.Open(virtualPath, FileMode.OpenOrCreate, access);
-        }
-
-        Stream IIOProvider.OpenFile(string filename)
-        {
-            return OpenFile(filename, FileAccess.ReadWrite);
+            return CurrentIOProvider.OpenFile(filename);
         }
 
         Stream IIOProvider.OpenFileReadOnly(string filename)
         {
-            return OpenFile(filename, FileAccess.Read);
+            if (CurrentIOProvider != null)
+            {
+                throw new NotSupportedException("Cannot open a file as a stream without an IO provider.");
+            }
+
+            var virtualPath = GetVirtualPath(filename);
+            if (!CurrentIOProvider.DirectoryExists(virtualPath))
+            {
+                CurrentIOProvider.CreateDirectory(virtualPath);
+            }
+            CurrentIOProvider.WriteAllBytes(virtualPath, (this as IIOProvider).ReadAllBytes(filename));
+
+            return CurrentIOProvider.OpenFileReadOnly(filename);
         }
 
         Stream IIOProvider.OpenFileWriteOnly(string filename)
         {
-            return OpenFile(filename, FileAccess.Write);
+            if (CurrentIOProvider != null)
+            {
+                throw new NotSupportedException("Cannot open a file as a stream without an IO provider.");
+            }
+
+            var virtualPath = GetVirtualPath(filename);
+            if (!CurrentIOProvider.DirectoryExists(virtualPath))
+            {
+                CurrentIOProvider.CreateDirectory(virtualPath);
+            }
+            CurrentIOProvider.WriteAllBytes(virtualPath, (this as IIOProvider).ReadAllBytes(filename));
+
+            return CurrentIOProvider.OpenFileWriteOnly(filename);
         }
         #endregion
